Refacotr Stage - Refactor code to follow design pattren DDD approach.
