--- src/OTB.JobOrdering.Kata.Application/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace OTB.JobOrdering.Kata.Application
-{
-    public class Class1
-    {
-    }
-}
