--- src/OTB.JobOrdering.Kata.Infrastructure/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace OTB.JobOrdering.Kata.Infrastructure
-{
-    public class Class1
-    {
-    }
-}
