+++ src/OTB.JobOrdering.Kata.Infrastructure/Extensions/IStringChecker.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace OTB.JobOrdering.Kata.Infrastructure.Extensions
+{
+    public interface IStringChecker
+    {
+        IEnumerable<string> SplitToLinesIgnoringEmpty(string jobSource);
+    }
+}
