+++ src/OTB.JobOrdering.Kata.Infrastructure/Repository/ISequenceRepository.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace OTB.JobOrdering.Kata.Infrastructure.Extensions
+{
+    public interface ISequenceRepository
+    {
+        string GetJobSequence();
+    }
+}
