+++ tests/OTB.JobOrdering.Kata.Application.Tests.Unit/JobOrderingTests.cs
@@ -1,3 +1,4 @@
+using OTB.JobOrdering.Kata.Application.Services;
 using System;
 using Xunit;
