+++ tests/OTB.JobOrdering.Kata.Application.Tests.Unit/OTB.JobOrdering.Kata.Application.Tests.Unit.csproj
@@ -19,4 +19,8 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\OTB.JobOrdering.Kata.Application\OTB.JobOrdering.Kata.Application.csproj" />
+  </ItemGroup>
+
 </Project>
